refactor(js): rename util files by category instead of function name
Steps:
- Rename `delay.js` → `timer.js`
- Rename `first-upper-case.js` → `string.js`
- Rename `get-current-year.js` → `date.js`
- Rename `svg-me.js` → `svg.js`
- Update all import paths in `scripts.js`, `card.js` and `copyright.js`

Reason:
- File names should represent the category, not the individual function
- Allows adding related functions to the same file in the future

References:
- Issue #104
