+++ assets/js/utils/string.js
